requirement.txt
astroid==2.1.0
Click==7.0
colorama==0.4.1
Django==2.2.5
django-crispy-forms==1.7.2
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==6.1.0
pylint==2.2.2
sqlparse==0.3.0
Werkzeug==0.15.4
wrapt==1.10.11
